+++ packages/design-tokens/src/js/themes/pureBlackDarkTheme/index.ts
@@ -1,1 +1,2 @@
 export { pureBlackDarkTheme } from './pureBlackDarkTheme';
+export { pureBlackDarkColorOverrides } from './colorOverrides';